Revert "[TASK] Make clearable button more accessible"
This reverts commit 75b558b6ee7a534462538d3356d9a4630ebc9201.

The proposed and merged solution decreases the overall usability in the
TYPO3 backend as users are now forced to press the <TAB> key twice to
actually switch the focus to another field.

One could also argue that the change that is reverted is not suitable
for the target group that mostly uses keyboard navigation as said users
were already able to clear fields previously.

Once we have a solution that satisfies both target groups, another
patch will be pushed.

Change-Id: I2cfea2aee81d2c67cba91eb5eb6d5675d6d999c6
Resolves: #102504
Reverts: #102372
Releases: main, 12.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81905
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Jochen Roth <rothjochen@gmail.com>