[TASK] Throw exception if `frontend.typoscript` is not available
Extension authors tend to use extbase way out of the
feel well habitat, for example using extbase in PSR-7
middlewares before the required frontend context stack
is there.

Avoiding a fatal php error has been mitigated with #102457
but still leaves the extension in the delusion of a working
context.

This change checks if the `frontent.typoscript` is not set
in the request and throwing a speaking runtime exception.

Resolves: #102511
Related: #102457
Releases: main
Change-Id: I3b4d71730958149f82c3256e81a369a694d30eeb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81955
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>