[BUGFIX] Use correct variable for preview links in element information
In the "References from this item" area of the "Element Information"
modal, there is s preview link for references of type page.
In #101276 a problem with the preview links was solved but another
one introduced so that the preview links were now always deactivated.
This patch now sets the correct variable in the fluid template.

Resolves: #102515
Related: #101276
Releases: main, 12.4
Change-Id: I4e129c98a067e015c7dd4e56af9eb006044bab83
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81959
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>