[BUGFIX] Wait for initialized FormEngine in `TemplateCest`
AC test reveal [1] that our code editor may be in a half-initialized
state before shutting down the test case, carrying over potential
JavaScript errors. To compensate this, `TemplateCest` now waits for the
target document and an additional 2 seconds to give the system a chance
to load everything properly.

[1] https://git.typo3.org/typo3/CI/cms/-/jobs/2778518

Resolves: #102526
Releases: main
Change-Id: I8c467cd1cdb0e253ca1a7917cfbc21020b9b6959
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81973
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>