[BUGFIX] Prevent endless loop in t3editor when using '/*' literal
When entering the sequence "/*" into any t3editor instance
(i.e. TypoScript editor in the backend), it puts the browser in an
endless loop.

The reason is that the method `stream.next()` declares to either return
a string value or `void` (=`undefined`), but the code checks for `null`.
The check is now fixed and another cases where a loose null comparison
was used is adapted as well. Note that a loose null comparison between
null and undefined return equality, but a strict comparison doesn't.

Resolves: #101893
Related: #101803
Releases: main, 12.4
Change-Id: If1c090daf13a6d730a2e5b96d46afab171fbcde2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82046
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>