[!!!][TASK] Raise minimal MySQL server version
Albeit creating a index for json-field array
values is not supported by MariaDB and SQLite,
the MySQL server version is raised to `8.0.17`
which adds support for a multi-value index.

Even if there is no ETA when `MariaDB` will
introduce a similar feature, this allows to
add a "json" based caching backend as a 3rd
party extension for MySQL DBMS, but have the
minimal version ensured.

Resolves: #102594
Related: #102518
Releases: main
Change-Id: I34a5688561b6e84b55f4c19f6abfe7dbf4619172
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82078
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>