[!!!][TASK] Avoid TSFE->applicationData
Property applicationData of TypoScriptFrontendController
is unfortunate: It is meant as an extension place to park
and communicate state using the just so convenient global
$GLOBALS['TSFE'] object.

Such things should of course be modeled differently, for
instance by adding own PSR-7 request attributes using a
custom middleware, or by firing own events for other
extensions to listen on.

In order to continue with our works on TSFE, the patch
removes that property.

The single use of that property in is ext:indexed_search:
It is used in combination with the crawler extension to
"force indexing". indexed_search now dispatches a proper
event instead.

Resolves: #102600
Releases: main
Change-Id: I2fdcbee8c196761670245cd90cb99414fc9beea7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82085
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>