[!!!][FEATURE] Add PSR-14 event for modifying GetData result
This introduces a new PSR-14 event, enabling
extension authors to modify the resolved "data",
just before `ContentObjectRenderer->getData()` is
about to return the result.

This replaces the previously available hook
$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['tslib/class.tslib_content.php']['getData']
which is therefore now removed.

The hook required implementations to implement
the `ContentObjectGetDataHookInterface`, which
is now deprecated as it's unused due to the
hook removal.

Resolves: #102614
Releases: main
Change-Id: I47cdfbcc03d7f820b9a8dd4a8989312c5261663e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82136
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>