[FEATURE] Introduce AsController attribute for auto configuration
This introduces a new PHP attribute `AsController`,
which serves as a drop-in replacement for the former
introduced `Controller` attribute (#99055), which is
now deprecated.

This is done to unify the naming of our attributes.

The old `Controller` attribute is still working using
class alias mapping.

For upwards compatibility, the #[AsController] attribute
will be backported to v12 with a more slim variant of
this patch.

Resolves: #102631
Related: #99055
Releases: main, 12.4
Change-Id: Ib0ff0d08660a89c4a5c6c14a327542ec89095743
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82152
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>