[TASK] Avoid TSFE->getContext() and TSFE->getRequestedId()
The context object is injectable singleton state - "Not good
but not terrible". There is no need to have a getter in
TypoScriptFrontendController for it. Refactor this. A test
case mocking can be simplified significantly with this.

The "originally requested id" can be retrieved from
the 'routing' Request attribute, there is not need to
carry a copy as TSFE property around.

Resolves: #102640
Releases: main
Change-Id: I816570339464ab38bb44b4e8df6773c2ed5c5a8e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82158
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Nikita Hovratov <nikita.h@live.de>