[TASK] Extract registration of controller actions into separate method
EU::configurePlugin tightly couples registration of controller actions
and TypoScript generation for a frontend rendering definition.
Furthermore, lib.contentElement is only defined in fluid_styled_content,
which makes it an indirect requirement.

This patch extracts the part for the controller action registration, so
it can be used independently of fluid_styled_content. In addition, this
new method expects the controller actions to have an array shape
already. A converter method ensures compatibility for both string and
array syntax.

A concrete use-case for internal usage would be Content Blocks, as it
would be possible to separately define a Content Block of type "Plugin"
and in addition register controller actions for it.

Resolves: #102643
Releases: main, 12.4
Change-Id: I74d84f54bdd399934b57b3e49e2209f62b5fda68
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82181
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>