[BUGFIX] Fix backend layout lookup in page info module
The page info module currently shows a valid page layout as invalid,
because the lookup of backend layouts still uses indexed array keys
instead of associative array keys which were introduced in #99739

This change fixes the problem by using associative array keys
to resolve backend layouts. Additionally, this change ensures,
that the `determinePageId` function in `BackendLayoutView`
does not try to resolve a page id, if the given array is empty.

Resolves: #102687
Related: #99739
Releases: main, 12.4
Change-Id: Iec74aec99c790c8841dbb2db537083503987f9d0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82198
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Nikita Hovratov <nikita.h@live.de>