[TASK] Update locales translation files
TYPO3 uses a composer package [1] to import
locales along with translations and a custom
script has been added to create and update
included translation files.

This change updates the language files with
the last updates and requires the package
with the current highest version as minimum.

Used command(s):

> composer req --dev \
    "sokil/php-isocodes-db-i18n":"^4.0.18"

> Build/Scripts/runTests.sh \
    -s checkIsoDatabase \
    -p 8.1

[1] sokil/php-isocodes-db-i18n

Resolves: #102747
Releases: main, 12.4
Change-Id: I4d6603484c4fca213c66f0df0feff13bc1011b29
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82306
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>