[TASK] Assert result based on symfony version in deprecated code tests
Symfony 7 removed deprecated code paths [1][2] and therefore started
using strict `in_array()` check for the `in` and `notIn` condition in
the expression language.

With #102746 the usage of symfony 7 has been allowed, which itself
introduces breaking changes. It is optional to upgrade to symfony 7
and therefore breaking changes has been respected and excludes from
the TYPO3 policies regarding breaking changes as it is not enforced.

This change modifies two tests to assert results based on the used
symfony version.

[1] https://github.com/symfony/symfony/pull/50867
[2] https://github.com/symfony/symfony/commit/7050bc82dd25c540958

Resolves: #102766
Related: #102746
Releases: 12.4
Change-Id: I0624f33ff270ba793054d0ff84810beee1a98c06
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82358
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>