[BUGFIX] Check all method for existence in `OpcodeCacheService`
The `\TYPO3\CMS\Core\Service\OpcodeCacheService->isClearable()`
method checks for existence of opcache function and if they are
listed in the `disable_functions` configuration.

It seems, that in some environments both methods are missing
and not listed in the disabled methods. Even if this sounds
odd when the opcache extension is installed but both methods are
not available, we add it to the check to avoid the missleading
error message:

	Call to undefined function
	TYPO3\CMS\Core\Service\opcache_reset()

Resolves: #102734
Releases: main, 12.4, 11.5
Change-Id: Icc17a276e8881f502a687db7ca57fc05ecb8981e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82384
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>