[BUGFIX] Prevent text wrap of title for file reference container
The title of a file reference container is currently wrapped to
multiple lines, depending on the length of the title and the
current screen resolution.

This change removes the fixed column width of the dt element.
Additionally, the debug information is now on the same line
as the title, which is the same behavior as for other inline
elements in TYPO3.

Resolves: #102066
Related: #99873
Releases: main, 12.4
Change-Id: Ic2b7d8b008a3dd4a95dce0f55228fd89e9b243af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82386
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>