[BUGFIX] Replace example in robots.txt
The robots.txt file currently defined in the site
config module on defining a static route is
outdated.

It excludes paths that do not exist anymore in
TYPO3 v12 in Composer.

Furthermore it does not suggest to specify a
sitemap.xml which is considered best practice
for SEO.

Resolves: #102771
Releases: main

Change-Id: I274d9823c73b73b84ff6d6650bf54419fa091057
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82365
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>