[TASK] Remove PSR-7 request attribute "target" for backend routes
Back in TYPO3 v7, the "target" request attribute for BE routes
was added with PSR-7, before Routing with PSR-15 and the
new Module API in TYPO3 v12 was added. It contained
the PHP class and method to be called as Request endpoint.

The Route object which is now used consistently, can now be
used and the target request attribute for backend requests
is removed again.

Resolves: #102781
Releases: main
Change-Id: I1e62e6c3395af9075f1433a1cb0547e4146b2807
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82400
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>