[TASK] Avoid invalid datetime database default in a test
With #102479 support for persisting datetime_immutable
object has been ensured and covered with a functional
test. However, using a empty string as default for a
not nullable datetime field is invalid for MariaDB and
MySQL which seemed to work until now in test executions.

Something has changed, but had no luck to find any traces
in the mariadb, mysql or corresponding docker hub changelogs
and commits explaining this.

The TYPO3 default TCA schema creates nullable fields with a
null default for `date, datetime and timestamp` anyway, and
therefore we now use that for the test case table structure.

It make sense to test a valid structure if working state is
expected.

Note:   It seems that core ci is not blocked by this yet,
        however local execution are broken. To verify with
        and without the change use following command(s):

    > Build/Scripts/runTests.sh -u
    > Build/Scripts/runTests.sh -s functional -d mysql \
        typo3/sysext/core/Tests/Functional/Database/

Resolves: #102807
Releated: #102479
Releases: main, 12.4
Change-Id: Ide32b3f9dc4b049ff5c3982997c85a0bb01e2fce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82423
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Markus Hofmann <typo3@calien.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Markus Hofmann <typo3@calien.de>