[TASK] Simplify extbase CacheService
Extbase CacheService has been extracted from the storage
backend in v11 with #94703, is still @internal and still
only used by the storage backend (fired in extbase bootstrap).

It has a weird construct that checks if a given table/uid
combination has a pid column. With this service being fired
from storage backends, which hard codes various pid calls
already, and with extbase repositories working on TCA tables
that have a pid column anyway, this is useless.

Additionally, there is code that checks if TSFE exists to
eventually clear page cache of current page uid. This is bogus
as well: Extbase is fired either as USER or USER_INT plugin.
In case of USER, a page cache entry can't exist, or the plugin
would not be rendered anyways. In case of USER_INT, the plugin
content is not cached, so it does not make sense to kill any
page cache entry.

Resolves: #102818
Related: #94703
Releases: main
Change-Id: Ifb90c2c6be7168e861983a204cc750dfed7ac5bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82436
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>