[TASK] Update container image versions
New core-testing-* images has been build and published in
the TYPO3 testing-infrastructure repository and pulled to
the core ci gitlab runner instances. [1][2]

This change modifies `Build/Scripts/runTests.sh` to set
new image version to ensure the usage for all changes
based on this change.

For information what changed in the image see the patches
in the testing-infrastrucure repository linked below - at
least for the phpXY images.

* core-testing-php81 => 2.10
* core-testing-php82 => 1.9
* core-testing-php83 => 1.10
* core-testing-apache24 => 1.3
* core-testing-nodejs18 => 1.2
* core-testing-nodejs18-chrome => 1.2

[1] https://git.typo3.org/typo3/CI/testing-infrastructure/-/commit/2fb328622d88d47afd3e8abc9bf6ff32f7ef3db6
[2] https://git.typo3.org/typo3/CI/testing-infrastructure/-/commit/e1ec0954b823acbab3f37100f8b6d0a03c31f0d8

Resolves: #102886
Releases: main, 12.4, 11.5
Change-Id: I3830a5f53b18e2170c9f973a66007dc3868d99dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82553
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>