[!!!][FEATURE] Add PSR-14 PackageInitializationEvent
This adds the PSR-14 PackageInitializationEvent. The Event
is being dispatched when packages are set up. TYPO3 Core
registers listeners, which execute corresponding tasks, such
as importing static extension data.

Note this in not about loading / unloading of extensions
in non-composer mode, but about setting up extensions
after being loaded in non-composer by extensionmanager,
or after being composer require'd (and thus auto-loaded)
in composer mode.

This allows to resolve dependencies to EXT:extensionmanager,
which is therefore now optional and can be omitted for composer
based installations. Activating packages can now be done by
calling the new PackageActivationService. This service does
also dispatch the new PSR-14 Event.

The former functionality in `InstallUtility->install()` is moved
to the new PackageActivationService. Therefore, the method now
only acts as wrapper around the new service, allowing listeners
to the AfterPackageActivationEvent to work as before.

However, following PSR-14 Events are removed from extension manager:

* AfterExtensionDatabaseContentHasBeenImportedEvent
* AfterExtensionFilesHaveBeenImportedEvent
* AfterExtensionSiteFilesHaveBeenImportedEvent
* AfterExtensionStaticDatabaseContentHasBeenImportedEvent

The information provided by those events can now be retrieved
by listeners to the PackageInitializationEvent using the event's
storage entries. Those entries are a PackageInitializationResult
object, providing the listener's result. By convention, the
identifier of the entries is the listener class name.

Using `before` and `after` in the listener registration it's
possible to place own listeners at a specific place, having
the necessary information available.

Additionally, the `InstallUtility->processExtensionSetup()` method
is removed. This method just triggered the corresponding tasks,
which can now be triggered by dispatching the new event.

With the new `CheckForImportRequirements` listener, TYPO3 checks
whether an extensions tries to import data, while required
components (e.g. EXT:impexp) are not installed. Instead of a
PHP Error, a warning is now logged in such case. If the package
activation is triggered via the `extension:setup` command, which
has been moved to EXT:core, the warning is also displayed on CLI.

Resolves: #102935
Releases: main
Change-Id: Icd340e85e64926d11c96c12461636f1606591cd1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82650
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>