[TASK] Stop exposing the internal AbstractFileFolder class
The abstract class `AbstractFileFolder` is broken by design as it
tries to abstract the `originalResource` property while this
property is relied on the have specific different types in
subclasses.

This is a pre-patch to get rid of this class and use clean native
type declarations in the subclasses.

As all affected classes are internal, this change is non-breaking.

Resolves: #102949
Related: #102879
Releases: main
Change-Id: I84e9d042fc186748e665bb9f37eddb30a339ad8a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82667
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>