[TASK] Clarify main tables must be on default connection
Configuring multiple database connections and then
moving 'arbitrary' tables to a connection other than
the "default" connection never worked well for the
main core table set.

We are using more and more queries that join data
accross tables, for instance with refindex and
categories. We have to continue with this to further
improve perfomance.

The patch adds an 'important.rst' to clarify the
situtation: Additional connections can still be
configured and used, but the core expects the "main"
table data set including all TCA tables to be on
the "default" connection.

Resolves: #102960
Releases: main
Change-Id: I6fbf4f85716c1bccb8741195056413701caa7ca4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82675
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>